added some updates

+++ escrow-vault/programs/escrow-vault/src/state.rs
@@ -9,6 +9,6 @@ pub struct Escrow {
     pub mint_a : Pubkey,
     pub mint_b : Pubkey,
     pub receive : u64,  // amount to receive from the maker
-    pub bump : u64
+    pub bump : u8
 }
